+++ examples/portal/src/app/app.component.ts
@@ -52,7 +52,6 @@ export class AppComponent implements OnInit {
                 hostParent: '#app-host-container',
                 hostClass: appHostClass,
                 routerPathPrefix: /\/app1|app4/, // '/app1',
-                selector: 'app1-root',
                 resourcePathPrefix: '/static/app1/',
                 preload: settings.app1.preload,
                 switchMode: settings.app1.switchMode,
@@ -73,7 +72,6 @@ export class AppComponent implements OnInit {
             {
                 name: 'app2',
                 hostParent: '#app-host-container',
-                selector: 'app2-root',
                 hostClass: appHostClass,
                 routerPathPrefix: '/app2',
                 resourcePathPrefix: '/static/app2/',