+++ packages/planet/src/helpers.spec.ts
@@ -5,7 +5,9 @@ import {
     getHTMLElement,
     getResourceFileName,
     buildResourceFilePath,
-    getScriptsAndStylesFullPaths
+    getScriptsAndStylesFullPaths,
+    getTagNameByTemplate,
+    createElementByTemplate
 } from './helpers';
 
 describe('helpers', () => {
@@ -122,6 +124,36 @@ describe('helpers', () => {
             getHTMLElement('.span');
             expect(querySelectorSpy).toHaveBeenCalledTimes(1);
             expect(querySelectorSpy).toHaveBeenCalledWith('.span');
+        });
+    });
+
+    describe(`createElementByTemplate`, () => {
+        it('should create html element success', () => {
+            const element = createElementByTemplate(`<app1-root class='app1-root-container'></app1-root>`);
+            expect(element.outerHTML).toBe(`<app1-root class="app1-root-container"></app1-root>`);
+        });
+
+        it('should create html element when template invalid', () => {
+            expect(() => {
+                createElementByTemplate(`app1-root`);
+            }).toThrowError(`invalid template 'app1-root'`);
+        });
+
+        it('should create html element when template is empty', () => {
+            const element = createElementByTemplate(``);
+            expect(element).toBe(null);
+        });
+    });
+
+    describe(`getTagNameByTemplate`, () => {
+        it('should get tagName is APP1-ROOT', () => {
+            const tagName = getTagNameByTemplate(`<app1-root></app1-root>`);
+            expect(tagName).toBe('APP1-ROOT');
+        });
+
+        it('should get tagName when template is empty', () => {
+            const tagName = getTagNameByTemplate(``);
+            expect(tagName).toBe(null);
         });
     });